+++ src/main/java/cz/muni/fi/pv260/team4/tictactoe/interfaces/IOProvider.java
@@ -4,27 +4,27 @@ import cz.muni.fi.pv260.team4.tictactoe.validators.InputValidator;
 
 public interface IOProvider {
     /**
-     * Read long value from the user.
+     * Read int value from the user.
      * <p>
      * If the validation fails, user will be prompted again for the value until validation succeeds.
      *
      * @param prompt    question to display
      * @param validator return empty for valid input or string error message
-     * @return read long value
+     * @return read int value
      */
-    long readLong(String prompt, InputValidator<Long> validator);
+    int readInt(String prompt, InputValidator<Integer> validator);
 
     /**
-     * Read long value from the user or use default on empty input.
+     * Read int value from the user or use default on empty input.
      * <p>
      * If the validation fails, user will be prompted again for the value until validation succeeds.
      *
      * @param prompt    question to display
      * @param defaultValue value to use on empty input
      * @param validator return empty for valid input or string error message
-     * @return read long value
+     * @return read int value
      */
-    long readLong(String prompt, Long defaultValue, InputValidator<Long> validator);
+    int readInt(String prompt, Integer defaultValue, InputValidator<Integer> validator);
 
     /**
      * Read String value from the user.
