+++ src/main/java/cz/muni/fi/pv260/team4/tictactoe/phase/SetupPhase.java
@@ -19,25 +19,25 @@ public final class SetupPhase implements GamePhase {
 
     @Override
     public GamePhase execute() {
-        var playerCount = ioProvider.readLong(
+        var playerCount = ioProvider.readInt(
                 "How many players",
                 Const.MINIMAL_PLAYER_COUNT,
                 new PlayerCountValidator(elementSupplier)
         );
 
-        var boardWidth = ioProvider.readLong(
+        var boardWidth = ioProvider.readInt(
                 "Width of the board",
                 Const.MINIMAL_BOARD_WIDTH,
                 new BoardWidthValidator()
         );
 
-        var boardHeight = ioProvider.readLong(
+        var boardHeight = ioProvider.readInt(
                 "Height of the board",
                 Const.MINIMAL_BOARD_HEIGHT,
                 new BoardHeightValidator()
         );
 
-        var lengthOfWinningSequence = ioProvider.readLong(
+        var lengthOfWinningSequence = ioProvider.readInt(
                 "Length of the winning sequence",
                 Const.MINIMAL_LENGTH_OF_WINNING_SEQUENCE,
                 new WinningSequenceLengthValidator(boardWidth, boardHeight)
